+++ Client/src/util/Team/TeamGamesSkeleton.js
@@ -4,8 +4,7 @@ import { makeStyles } from '@material-ui/core/styles';
 // Components
 import DataWrapper from '../../components/DataWrapper';
 import TeamHeader from '../../components/Team/TeamHeader';
-import TeamGames from '../../components/Team/TeamGames';
-import TeamScouting from '../../components/Team/TeamScouting';
+import TeamGamesWrapper from '../../components/Team/TeamGamesWrapper';
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -17,18 +16,15 @@ export default function TeamGamesSkeleton({ info, scouting, games }) {
     const classes = useStyles();
 
     let headerComponent = (<TeamHeader info={info} type='Games' />);
-    let scoutingComponent = (<TeamScouting scouting={scouting} />);
-    let gamesComponent = (<TeamGames games={games} />);
+    let gamesWrapperComponent = (<TeamGamesWrapper info={info} scouting={scouting} games={games} />);
 
-    let headerEmpty = "There is no Information logged for this Tournament.";
-    let scoutingEmpty = "There is no Scouting information for this Team."
-    let gamesEmpty = "There are no Games logged for this Tournament.";
+    let headerEmpty = "There is no Information logged for this Team.";
+    let gamesEmpty = "There are no Games logged for this Team.";
 
     return (
         <div className={classes.root}>
             <DataWrapper data={info} component={headerComponent} emptyMessage={headerEmpty} />
-            <DataWrapper data={scouting} component={scoutingComponent} emptyMessage={scoutingEmpty} />
-            <DataWrapper data={games} component={gamesComponent} emptyMessage={gamesEmpty} />
+            <DataWrapper data={games && scouting} component={gamesWrapperComponent} emptyMessage={gamesEmpty} />
         </div>
     )
 }