+++ Client/src/util/Team/TeamStatsSkeleton.js
@@ -16,9 +16,9 @@ export default function TeamStatsSkeleton({ info, stats }) {
     const classes = useStyles();
 
     let headerComponent = (<TeamHeader info={info} type='Stats' />);
-    let statsComponent = (<TeamStats stats={stats} />);
+    let statsComponent = (<TeamStats info={info} stats={stats} />);
 
-    let headerEmpty = "There is no Information logged for this Tournament.";
+    let headerEmpty = "There is no Information logged for this Team.";
     let statsEmpty = "There are no Stats logged for this Team."
 
     return (