+++ functions/profileData.js
@@ -82,6 +82,16 @@ function getProfileInfo(pPId) {
                         if ('ActiveTeamHId' in profileInfoJson) {
                             profileInfoJson['ActiveTeamName'] = await Team.getName(profileInfoJson['ActiveTeamHId']);
                         }
+                        // Add Season List
+                        let gameLogJson = (await dynamoDb.getItem('Profile', 'ProfilePId', pPId, ['GameLog']))['GameLog'];
+                        if (gameLogJson != null) {
+                            profileInfoJson['SeasonList'] = await helper.getSeasonItems(Object.keys(gameLogJson));
+                        }
+                        // Add Tournament List
+                        let statsLogJson = (await dynamoDb.getItem('Profile', 'ProfilePId', pPId, ['StatsLog']))['StatsLog'];
+                        if (statsLogJson != null) {
+                            profileInfoJson['TournamentList'] = await helper.getTourneyItems(Object.keys(statsLogJson));
+                        }
                         cache.set(cacheKey, JSON.stringify(profileInfoJson, null, 2));
                         resolve(profileInfoJson);
                     }