+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Overall
+* vA.7 - Client 0.0.9, Server 0.0.7
 * vA.6 - Client 0.0.8, Server 0.0.6
 * vA.5 - Client 0.0.7, Server 0.0.5
 * vA.4 - Client 0.0.6, Server 0.0.4
@@ -14,10 +15,12 @@
 * 1.0.0 will be the first production website
 
 # Client
++ 0.0.9
+    + Team pages for Games and Stats done
 + 0.0.8
     + Skeletons finished for Teams, Profiles, and Matches
 + 0.0.7
-    + Cleaned up Client to accommodate server's new error handling.
+    + Cleaned up Client to accommodate server's new error handling
     + Implemented Skeletons for Team
 + 0.0.6
     + Tournament pages completed
@@ -36,6 +39,8 @@
     + First Github commit for collaboration purposes.
 
 # Server
++ 0.0.7
+    + Modified teams API slightly to conform to Client change.
 + 0.0.6
     + Modified additional API routes for latest season and tournaments
 + 0.0.5